requirements.txt
ipython==5.4.1
tqdm
numpy
nose
opentrons
PyYAML